refactor(writers): share the markdown-family link and code-span helpers
